// src/plugins/cms/ui/generated-admin-types.ts
export type Maybe<T> = T;
export type InputMaybe<T> = T;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  DateTime: { input: any; output: any; }
  JSON: { input: any; output: any; }
  Money: { input: number; output: number; }
  Upload: { input: any; output: any; }
};
// ...
export type CreatePageInput = {
  content: Scalars['String']['input'];
  featured_image_id?: InputMaybe<Scalars['ID']['input']>;
  meta_description: Scalars['String']['input'];
  meta_keyword: Scalars['String']['input'];
  meta_title: Scalars['String']['input'];
  nav: Scalars['String']['input'];
  parent_id?: InputMaybe<Scalars['ID']['input']>;
  show_in_nav: Scalars['Boolean']['input'];
  slug: Scalars['String']['input'];
  sort_order: Scalars['Int']['input'];
  status: Scalars['String']['input'];
  title: Scalars['String']['input'];
};
// ...
export type DeletionResponse = {
  __typename?: 'DeletionResponse';
  message?: Maybe<Scalars['String']['output']>;
  result: DeletionResult;
};

export enum DeletionResult {
  /** The entity was successfully deleted */
  DELETED = 'DELETED',
  /** Deletion did not take place, reason given in message */
  NOT_DELETED = 'NOT_DELETED'
}
// ...
export type UpdatePageInput = {
  content?: InputMaybe<Scalars['String']['input']>;
  featured_image_id?: InputMaybe<Scalars['ID']['input']>;
  id: Scalars['ID']['input'];
  meta_description?: InputMaybe<Scalars['String']['input']>;
  meta_keyword?: InputMaybe<Scalars['String']['input']>;
  meta_title?: InputMaybe<Scalars['String']['input']>;
  nav?: InputMaybe<Scalars['String']['input']>;
  parent_id?: InputMaybe<Scalars['ID']['input']>;
  show_in_nav?: InputMaybe<Scalars['Boolean']['input']>;
  slug?: InputMaybe<Scalars['String']['input']>;
  sort_order?: InputMaybe<Scalars['Int']['input']>;
  status?: InputMaybe<Scalars['String']['input']>;
  title?: InputMaybe<Scalars['String']['input']>;
};